Move interface expand/flatten to helper package
