+++ linode/instance/expand.go
@@ -42,13 +42,3 @@ func expandInstanceConfigDevice(m map[string]interface{}) *linodego.InstanceConf
 	}
 	return dev
 }
-
-func expandConfigInterface(i map[string]interface{}) linodego.InstanceConfigInterfaceCreateOptions {
-	result := linodego.InstanceConfigInterfaceCreateOptions{}
-
-	result.Label = i["label"].(string)
-	result.Purpose = linodego.ConfigInterfacePurpose(i["purpose"].(string))
-	result.IPAMAddress = i["ipam_address"].(string)
-
-	return result
-}
